+++ ResourceGenerator/M365DSCResourceGenerator.psm1
@@ -176,13 +176,15 @@ function New-M365DSCResource
 
         $global:ComplexList=@()
         $cimClasses = Get-Microsoft365DSCModuleCimClass -ResourceName $ResourceName
+        $global:searchedEntity = $selectedODataType
         $typeProperties = Get-TypeProperties `
             -CmdletDefinition $cmdletDefinition `
             -Entity $selectedODataType `
             -IncludeNavigationProperties $IncludeNavigationProperties `
             -CimClasses $cimClasses `
             -Workload $Workload
-        $global:ComplexList=$null
+        $global:ComplexList = $null
+        $global:searchedEntity = $null
         [Hashtable[]]$parameterInformation = Get-ParameterBlockInformation `
             -Properties $typeProperties `
             -DefaultParameterSetProperties $defaultParameterSetProperties
@@ -1247,10 +1249,8 @@ function Get-TypeProperties
             $rawProperties=$entityType.Property
             foreach ($property in $rawProperties)
             {
-
-
                 $IsRootProperty=$false
-                if (($entityType.BaseType -eq "graph.Entity") -or ($entityType.Name -eq "entity") -or $isAbstract)
+                if (($entityType.BaseType -eq "graph.Entity") -or ($entityType.Name -eq "entity") -or ($isAbstract -and $entityType.Name -eq $global:searchedEntity))
                 {
                     $IsRootProperty=$true
                 }