+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
   * IntuneDeviceConfigurationPkcsCertificatePolicyWindows10
     * Initial release
+ * DRG
+  * Fix issue with abstract type and additionalProperties
 
 # 1.23.426.3