+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 # UNRELEASED
 
+
  * IntuneDeviceConfigurationTrustedCertificatePolicyWindows10
   * Initial release
+ * IntuneDeviceConfigurationPkcsCertificatePolicyWindows10
+   * Initial release
+ * DRG
+  * Fix issue with abstract type and additionalProperties
 
 # 1.23.426.3