+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * IntuneDeviceConfigurationPkcsCertificatePolicyWindows10
   * Initial release
+* IntuneDeviceConfigurationVpnPolicyWindows10
+  * Initial release
 * MISC
   * M365DSCUtil: Fixed an issue on function Get-M365DSCExportContentForResource if ConnectionMode was set to anything but "Credentials*"
  * DRG