+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   * Initial release
 * IntuneDeviceConfigurationPkcsCertificatePolicyWindows10
   * Initial release
+ * IntuneDeviceConfigurationSecureAssessmentPolicyWindows10
+  * Initial release
 * IntuneDeviceConfigurationTrustedCertificatePolicyWindows10
   * Initial release
 * IntuneDeviceConfigurationVpnPolicyWindows10