+++ CHANGELOG.md
@@ -2,11 +2,19 @@
 
 # UNRELEASED
 
-* IntuneDeviceConfigurationVpnPolicyWindows10
+* IntuneDeviceConfigurationImportedPfxCertificatePolicyWindows10
   * Initial release
 * IntuneDeviceConfigurationPkcsCertificatePolicyWindows10
   * Initial release
 * IntuneDeviceConfigurationScepCertificatePolicyWindows10
+  * Initial release
+ * IntuneDeviceConfigurationSecureAssessmentPolicyWindows10
+  * Initial release
+* IntuneDeviceConfigurationSharedMultiDevicePolicyWindows10
+  * Initial release
+* IntuneDeviceConfigurationTrustedCertificatePolicyWindows10
+  * Initial release
+* IntuneDeviceConfigurationVpnPolicyWindows10
   * Initial release
 * DRG
   * Fix issue with abstract type and additionalProperties